remove dependency on react GA for widget
